# Ensure no validation when active_validation is False