--- a/requirements.txt
+++ b/requirements.txt
@@ -5,7 +5,7 @@ numpy
 jsonschema
 jsonref
 pyelftools
-
+fastjsonschema==2.16.1
 # Tests
 pytest