--- a/setup.cfg
+++ b/setup.cfg
@@ -35,9 +35,9 @@ install_requires =
     numpy>=1.13
     zhinst-core>=23.06
     zhinst-utils>=0.3.1
-    jsonschema>=3.2.0
     jsonref>=0.2
     typing_extensions>=4.1.1
+    fastjsonschema==2.16.1
     pyelftools>=0.29
 
 include_package_data = True