Description:

This PR addresses performance issues in #269

JSON schema validation decreases the performance of the function and should be bypassed if
the user does not want the validation and requires performance.

Changed to `fastjsonschema` which is much faster.


---

I ran the example in the issue #269 locally and results are for `active_validation=True`:

Before:

- with JSON Schema validation: ~10 ms
- without validation: 0.01 - 0.04 ms

After:

- with JSON Schema validation: ~0.01 - 0.04 ms
- without validation: 0.01 - 0.04 ms

Fixes issue: #269 

Checklist:

- [ ] Add tests for the change to show correct behavior.
- [ ] Add or update relevant docs, code and examples.
- [ ] Update CHANGELOG.rst with relevant information and add the issue number.
- [ ] Add .. versionchanged:: where necessary.