@markheik 2023-10-19
I am still unsure if we want to acknowledge `active_validation` value here, but the default should be `True`.