@markheik 2023-10-20
I will close this in favor of #271  As this is out of scope for #269 . 
Will revisit the overall performance improvements later.